requirements.txt
python-telegram-bot==12.7
selenium==3.141.0
pillow==7.2.0
requests==2.24.0